requirements.txt
# MedSupply AI - Python Dependencies
# Python 3.8+ required

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=3.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning (Optional)
tensorflow>=2.13.0
keras>=2.13.0

# Time Series Analysis
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.28.0

# Data Processing
scikit-fuzzy>=0.4.2
openpyxl>=3.1.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Scheduling
apscheduler>=3.10.0

# Model Persistence
joblib>=1.3.0

# Google Sheets Integration (Optional)
gspread>=5.12.0
google-auth>=2.23.0

# Environment Config
python-dotenv>=1.0.0

# HTTP Client (dashboard -> API)
requests>=2.31.0

# Utilities (stdlib – listed for documentation only)
# warnings, datetime, logging, uuid, json, pathlib

# Development (Optional)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

